requirements.txt
streamlit
yt-dlp
gallery-dl
openai-whisper
pandas
